feat(custody): /kotoba/key/1 custodian protocol core (ADR-sealed-cold-tier R3b)
Transport-agnostic core of the threshold key-release protocol:

- KeyShareRequest/KeyShareResponse wire types (requester carries an ephemeral
  X25519 pubkey; the custodian re-wraps its share to it)
- handle_key_share_request: authorize() FIRST (CACAO + purpose + nonce +
  receipt, injected by the server layer) → open at-rest share → re-wrap to
  requester. 'no receipt, no key' is control-flow-enforced: a denied request
  never touches share material; the authorize hook fires before any release
- combine_granted: requester opens t re-wrapped shares and recombines locally
- authorization injected as a closure → kotoba-custody stays a leaf crate
  (same seam discipline as the R2c import check)

Split lib.rs into shares (R3a) + protocol (R3b). 14 tests green (8 shares +
6 protocol: roundtrip, denial-releases-nothing, authorize-runs-first,
in-flight-share-readable-only-by-requester, sub-threshold, serde). clippy clean.
libp2p Behaviour shell deferred (kotoba-turn #102 pattern).
