+++ deps.toml
@@ -152,7 +152,7 @@ notes = "AES-256-GCM seal/open with OsRng nonce; Zeroizing<Vec<u8>> plaintext on
 [subdirs."crates/kotoba-custody"]
 role = "key-custody"
 tier = "alpha"
-notes = "R3a share plane (2026-06-11, ADR-sealed-cold-tier R3): t-of-N Shamir custody of KOTOBA_BLOCK_KEY. split_key(key,t,custodians)→CustodianShare{index, recipient_did, sealed_share=HPKE(ephemeral_pk||nonce||AES-256-GCM via kotoba-crypto), commitment=sha256(share), threshold} — Shamir GF(2^8) via sharks crate (audited, not hand-rolled); open_share verifies commitment (corrupt/substituted share detected); combine_key(t,shares)→Zeroizing<[u8;32]>; t-1 shares = information-theoretic nothing. 8 tests (3-of-5 roundtrip, insufficient-shares, wrong-recipient HPKE, commitment mismatch, randomized dealing, cross-dealing garbage, bad threshold, serde). R3b = /kotoba/key/1 custodian protocol (CACAO+purpose+receipt per share release); R3c = Feldman VSS + MLS-epoch rotation; R3d = MishmarBondEscrow bonds + warrants."
+notes = "R3a share plane (2026-06-11, ADR-sealed-cold-tier R3): t-of-N Shamir custody of KOTOBA_BLOCK_KEY. split_key(key,t,custodians)→CustodianShare{index, recipient_did, sealed_share=HPKE(ephemeral_pk||nonce||AES-256-GCM via kotoba-crypto), commitment=sha256(share), threshold} — Shamir GF(2^8) via sharks crate (audited, not hand-rolled); open_share verifies commitment (corrupt/substituted share detected); combine_key(t,shares)→Zeroizing<[u8;32]>; t-1 shares = information-theoretic nothing. 8 tests (3-of-5 roundtrip, insufficient-shares, wrong-recipient HPKE, commitment mismatch, randomized dealing, cross-dealing garbage, bad threshold, serde). R3b core (2026-06-11): protocol.rs — KeyShareRequest{graph,cacao_b64,purpose,nonce,requester_x25519_pk}→KeyShareResponse::{Granted(GrantedShare re-wrapped to requester),Denied}; handle_key_share_request calls injected authorize() FIRST (CACAO+purpose+nonce+receipt) then opens+re-wraps share — \"no receipt, no key\" control-flow-enforced; combine_granted(t,grants,requester_sk) recombines client-side; 6 protocol tests (roundtrip, denial-releases-nothing, authorize-runs-first, in-flight-share-requester-only, sub-threshold, serde). libp2p request-response Behaviour = deferred shell (kotoba-turn #102 pattern). R3c = Feldman VSS + MLS-epoch rotation; R3d = MishmarBondEscrow bonds + warrants."
 
 [subdirs."crates/kotoba-store"]
 role = "block-store"
